pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Discord ticket bot for carry services"
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.0.0",
    "python-dotenv>=1.0.0",
    "motor>=3.3.0",
    "pymongo>=4.6.0",
    "asyncio>=3.4.3",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"